requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
pydantic==2.9.2
Pillow==11.0.0
ImageHash==4.3.1
numpy==1.26.4
matplotlib==3.9.2
ultralytics==8.3.0
pyyaml==6.0.2
